[BE] Department, Member dto entity 변환 수정 및 테스트 코드 변경

+++ src/main/java/com/kbe5/rento/domain/member/service/MemberService.java
@@ -1,16 +1,16 @@
 package com.kbe5.rento.domain.member.service;
 
 import com.kbe5.rento.common.exception.DomainException;
+import com.kbe5.rento.common.exception.ErrorType;
 import com.kbe5.rento.domain.company.entity.Company;
 import com.kbe5.rento.domain.company.repository.CompanyRepository;
 import com.kbe5.rento.domain.department.entity.Department;
 import com.kbe5.rento.domain.department.repository.DepartmentRepository;
-import com.kbe5.rento.domain.member.dto.request.MemberRegisterRequest;
+import com.kbe5.rento.domain.manager.entity.Manager;
 import com.kbe5.rento.domain.member.dto.request.MemberUpdateRequest;
 import com.kbe5.rento.domain.member.dto.response.MemberInfoResponse;
 import com.kbe5.rento.domain.member.entity.Member;
 import com.kbe5.rento.domain.member.repository.MemberRepository;
-import com.kbe5.rento.common.exception.ErrorType;
 import lombok.RequiredArgsConstructor;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
@@ -28,59 +28,65 @@ public class MemberService {
     private final CompanyRepository companyRepository;
 
     @Transactional
-    public void register(MemberRegisterRequest request) {
-        Department department = departmentRepository.findById(request.departmentId())
-                .orElseThrow(() -> new DomainException(ErrorType.DEPARTMENT_NOT_FOUND));
+    public Member register(Member member, Long departmentId) {
+        validateDuplicatePhoneNumber(member.getPhoneNumber(), member.getCompanyCode());
 
-        Company company = companyRepository.findByCompanyCode(request.companyCode())
-                .orElseThrow(() -> new DomainException(ErrorType.COMPANY_NOT_FOUND));
+        member.encodePassword(passwordEncoder);
 
-        String password = passwordEncoder.encode(request.password());
+        member.assignCompany(companyRepository.findByCompanyCode(member.getCompanyCode()).orElseThrow(
+                () -> new DomainException(ErrorType.COMPANY_NOT_FOUND))
+        );
+        member.assignDepartment(departmentRepository.findById(departmentId).orElseThrow(
+                () -> new DomainException(ErrorType.DEPARTMENT_NOT_FOUND))
+        );
 
-        validateDuplicatePhoneNumber(request.phoneNumber(), request.companyCode());
-
-        Member member =  MemberRegisterRequest.toEntity(request, password, department, company);
-
-        memberRepository.save(member);
+        return memberRepository.save(member);
     }
 
     @Transactional
-    public void update(MemberUpdateRequest request, Long memberId) {
+    public MemberInfoResponse update(Manager manager, MemberUpdateRequest request, Long memberId) {
         Member member = memberRepository.findById(memberId)
                 .orElseThrow(() -> new DomainException(ErrorType.MEMBER_NOT_FOUND));
 
         Department department = departmentRepository.findById(request.departmentId())
                 .orElseThrow(() -> new DomainException(ErrorType.DEPARTMENT_NOT_FOUND));
+
+        if(!manager.getCompany().getId().equals(member.getCompany().getId())) {
+            throw new DomainException(ErrorType.NOT_AUTHORIZED);
+        }
 
         validateDuplicatePhoneNumber(request.phoneNumber(), request.companyCode());
 
         member.update(request.name(), request.email(), request.getPosition(), request.loginId(), request.phoneNumber(), department);
+
+        return MemberInfoResponse.from(member);
     }
 
     @Transactional
-    public void delete(Long memberId) {
+    public void delete(Manager manager, Long memberId) {
         Member member = memberRepository.findById(memberId)
                 .orElseThrow(() -> new DomainException(ErrorType.MEMBER_NOT_FOUND));
+
+        if(!manager.getCompany().getId().equals(member.getCompany().getId())) {
+            throw new DomainException(ErrorType.NOT_AUTHORIZED);
+        }
 
         memberRepository.delete(member);
     }
 
     @Transactional(readOnly = true)
-    public List<MemberInfoResponse> getMemberList(String companyCode) {
+    public List<Member> getMemberList(String companyCode) {
         Company company = companyRepository.findByCompanyCode(companyCode)
                 .orElseThrow(() -> new DomainException(ErrorType.COMPANY_NOT_FOUND));
 
-        return memberRepository.findAllByCompanyId(company.getId()).stream()
-                .map(MemberInfoResponse::from)
-                .toList();
+        return memberRepository.findAllByCompanyId(company.getId());
+
     }
 
     @Transactional(readOnly = true)
-    public MemberInfoResponse getMember(Long memberId) {
-        Member member =  memberRepository.findById(memberId)
+    public Member getMember(Long memberId) {
+        return memberRepository.findById(memberId)
                 .orElseThrow(() -> new DomainException(ErrorType.MEMBER_NOT_FOUND));
-
-        return MemberInfoResponse.from(member);
     }
 
     private void validateDuplicatePhoneNumber(String phoneNumber, String companyCode) {
